add format_and_print_results_table
