+++ examples/helpers.py
@@ -6,9 +6,13 @@ from pathlib import Path
 from typing import TYPE_CHECKING
 
 import requests
+from tabulate import tabulate
 
 if TYPE_CHECKING:
     from collections.abc import Collection
+
+    import pandas as pd
+
 logger = logging.getLogger(__name__)
 
 BYTES_IN_MB = 1024 * 1024
@@ -73,3 +77,61 @@ def download_zenodo_data(
         else:
             logger.info(f"File {dst_fpath} already exists. Skipping download.")
         filepaths.append(dst_fpath)
+
+
+def format_and_print_results_table(
+    results_per_test_ref_df: pd.DataFrame, *, print_small_table: bool = False
+) -> pd.DataFrame:
+    key_results_df = results_per_test_ref_df[
+        [
+            "test_wtg",
+            "ref",
+            "uplift_frc",
+            "unc_one_sigma_frc",
+            "uplift_p95_frc",
+            "uplift_p5_frc",
+            "pp_valid_hours_pre",
+            "pp_valid_hours_post",
+            "mean_power_post",
+        ]
+    ]
+
+    def _convert_frc_cols_to_pct(input_df: pd.DataFrame, dp: int = 1) -> pd.DataFrame:
+        for i, col in enumerate(x for x in input_df.columns if x.endswith("_frc")):
+            if i == 0:
+                output_df = input_df.assign(**{col: (input_df[col] * 100).round(dp).astype(str) + "%"})
+            else:
+                output_df = output_df.assign(**{col: (input_df[col] * 100).round(dp).astype(str) + "%"})
+            output_df = output_df.rename(columns={col: col.replace("_frc", "_pct")})
+        return output_df
+
+    print_df = _convert_frc_cols_to_pct(key_results_df).rename(
+        columns={
+            "test_wtg": "turbine",
+            "ref": "reference",
+            "uplift_pct": "energy uplift",
+            "unc_one_sigma_pct": "uplift uncertainty",
+            "uplift_p95_pct": "uplift P95",
+            "uplift_p5_pct": "uplift P5",
+            "pp_valid_hours_pre": "valid hours toggle off",
+            "pp_valid_hours_post": "valid hours toggle on",
+            "mean_power_post": "mean power toggle on",
+        }
+    )
+    print_df["mean power toggle on"] = print_df["mean power toggle on"].round(0).astype("int64")
+    print_df_for_tabulate = (
+        print_df[["turbine", "reference", "energy uplift", "uplift P95", "uplift P5", "valid hours toggle on"]]
+        if print_small_table
+        else print_df
+    )
+    results_table = tabulate(
+        print_df_for_tabulate,
+        headers="keys",
+        tablefmt="outline",
+        floatfmt=".1f",
+        numalign="center",
+        stralign="center",
+        showindex=False,
+    )
+    print(results_table)
+    return print_df
